# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for Jingliange.",
        "title": "Jingliange Server API",
        "termsOfService": "TO IMPLEMENT",
        "contact": {},
        "version": "1.0"
    },
    "host": "49.234.22.169:8000",
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "description": "简单的用户认证，基于微信小程序的用户ID，返回JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "用户认证",
                "parameters": [
                    {
                        "description": "认证参数",
                        "name": "auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{\"token\":\"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\",\"user_id\":\"user123\"}}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"invalid params\",\"data\":\"Invalid input data\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"internal server error\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/donation/createDonation": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "用户支付成功后创建捐款记录（需要JWT认证）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Donation"
                ],
                "summary": "创建捐款记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "捐款信息",
                        "name": "donation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DonationCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":\"Donation created successfully\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"invalid params\",\"data\":\"Invalid input data\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "401": {
                        "description": "{\"code\":401,\"msg\":\"unauthorized\",\"data\":\"Token required\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"internal server error\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/donation/getDonationList": {
            "post": {
                "description": "获取捐款记录列表，支持年份、时间段、昵称筛选，支持按时间或金额排序，支持分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Donation"
                ],
                "summary": "获取功德榜列表",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DonationQueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":[{\"id\":1,\"user_id\":\"user123\",\"donor_name\":\"善心人士\",\"amount\":100.00,\"donate_time\":\"2025-07-08T10:00:00Z\",\"is_visible\":1,\"message\":\"祝愿净莲阁越来越好\"}]}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Donation"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"invalid params\",\"data\":\"Invalid input data\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"internal server error\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/donation/getDonationStats": {
            "post": {
                "description": "获取指定时间段的捐款统计信息，包括总金额和总人次",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Donation"
                ],
                "summary": "获取捐款统计",
                "parameters": [
                    {
                        "description": "统计参数",
                        "name": "stats",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DonationStatsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{\"total_amount\":5000.00,\"total_count\":50}}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.DonationStats"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"invalid params\",\"data\":\"Invalid input data\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"internal server error\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.AuthRequest": {
            "type": "object",
            "properties": {
                "user_id": {
                    "description": "用户ID（来自微信小程序）",
                    "type": "string",
                    "example": "user123"
                }
            }
        },
...
        "model.Donation": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "捐款金额（必填）",
                    "type": "number",
                    "example": 100
                },
                "create_time": {
                    "type": "string",
                    "example": "2025-07-08T10:00:00Z"
                },
                "donate_time": {
                    "description": "捐款时间（必填）",
                    "type": "string",
                    "example": "2025-07-08T10:00:00Z"
                },
                "donor_name": {
                    "description": "捐款人昵称（必填）",
                    "type": "string",
                    "example": "善心人士"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_visible": {
                    "description": "是否显示在榜单：0隐藏，1显示",
                    "type": "integer",
                    "example": 1
                },
                "message": {
                    "description": "留言",
                    "type": "string",
                    "example": "祝愿净莲阁越来越好"
                },
                "remarks": {
                    "description": "备注",
                    "type": "string",
                    "example": ""
                },
                "update_time": {
                    "type": "string",
                    "example": "2025-07-08T10:00:00Z"
                },
                "user_id": {
                    "description": "用户ID（必填）",
                    "type": "string",
                    "example": "user123"
                }
            }
        },
        "model.DonationCreateRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "捐款金额",
                    "type": "number",
                    "example": 100
                },
                "donor_name": {
                    "description": "捐款人昵称",
                    "type": "string",
                    "example": "善心人士"
                },
                "message": {
                    "description": "留言",
                    "type": "string",
                    "example": "祝愿净莲阁越来越好"
                }
            }
        },
        "model.DonationQueryRequest": {
            "type": "object",
            "properties": {
                "donor_name": {
                    "description": "捐款人昵称搜索（模糊匹配）",
                    "type": "string",
                    "example": "善心"
                },
                "page_number": {
                    "description": "页码，从零开始",
                    "type": "integer",
                    "example": 0
                },
                "page_size": {
                    "description": "每页数量",
                    "type": "integer",
                    "example": 10
                },
                "period": {
                    "description": "时间范围：all全年，first上半年，second下半年",
                    "type": "string",
                    "example": "all"
                },
                "sort_by": {
                    "description": "排序方式：time时间，amount金额",
                    "type": "string",
                    "example": "time"
                },
                "sort_order": {
                    "description": "排序顺序：desc降序，asc升序",
                    "type": "string",
                    "example": "desc"
                },
                "year": {
                    "description": "年份，0表示不限制",
                    "type": "integer",
                    "example": 2025
                }
            }
        },
        "model.DonationStats": {
            "type": "object",
            "properties": {
                "total_amount": {
                    "description": "总捐款金额",
                    "type": "number",
                    "example": 5000
                },
                "total_count": {
                    "description": "总捐款人次",
                    "type": "integer",
                    "example": 50
                }
            }
        },
        "model.DonationStatsRequest": {
            "type": "object",
            "properties": {
                "period": {
                    "description": "时间范围：all全年，first上半年，second下半年",
                    "type": "string",
                    "example": "all"
                },
                "year": {
                    "description": "年份，0表示不限制",
                    "type": "integer",
                    "example": 2025
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}